requirements.txt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.4
factory-boy==2.11.1
Faker==1.0.0
idna==2.8
mock==2.0.0
mysqlclient==1.3.14
nose==1.3.7
pbr==5.1.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
